Fix rake asset:precompile
When compiling assets the jasmine gem isn't defined because we
specifically activate only the 'assets' gem group.

The routes file tries to avoid mixing in the jasmine routes in this
case, but this initializer causes the JasmineRails module to be always
defined even when the jasmine gem isn't loaded.

This changes the initializer to avoid defining the JasminRails module if
it's not already defined.