Display Finder Organisation
In order for the user to be able to navigate back to the Org, this
commit displays a link back to the Finder's tagged Org. This commit also
styles it to look like the metadata in the rest of GOV.UK and adds a
CSS reset to fix some browser stylesheet margins going on. I also fixed
a spacing issue this introduced on the Filter Results button.