Always parse radio facets as Arrays
In order to link back to the pre filtered Finders, we need to have a
started way of linking backing to select and radio facets. To do this,
we turn the radio facet values into an array and take the first value
when we check it against the allowed_values. This also means we should
remove the test which was safeguarding against this.