Split result count + results into separate musache templates
Split result count and results into two separate mustache templates so that screen readers will only announce the result count, not read out the whole list of results.

This behaviour requires two templates because the aria-live region must wrap the changing bit, not be part of it, so must exist outside of the mustache, which means separating the aria-live bits and the not aria-live bits so they can be wrapped with HTML that remains on the page.