Add any_filters_applied filter
User research revealed that users were being conufsed by the big number at the top of the page when they initially landed on the finder, thinking it meant they had already completed a search when they hadn't.

This commit adds in a flag to the .to_hash of the result_set which tells us if any filters have been applied. Based on the value of this filter we can then show or not show the result count.