Order by last_update when not searching by keyword
In order to match the finder-api behaviour that rummager is taking over,
we want to order results when filtering by the date they were last
updated.

This last_update field, once added to rummager in the schema and
permitted sort fields, and provided by the publishing applications, will
allow this behaviour to be maintained.