Introduce email alert subscription workflow
This is the basis of the workflow for fetching topic urls from govuk_delivery api. It would be nice to remove the reliance on the Finder so heavily however as an MVP this should be sufficient. We may require the finder later on.

We have tried to decouple the EmailAlertSignupAPI wrapper from the Finder as much as possible so if we do wish to change this later it should be much simpler.