Parse six digit dates correctly
With the new Date Facet added in 3376192, Ruby's `Date.parse` method was
parsing six digit dates the wrong way round, so 25/09/14 would be parsed
as 14th September 2025. This commit fixes that by checking if it's a 6
digit date and then building the correct Date, if it doesn't match it
hands the input to `Date.parse`.