Fix for Facet Changes
After the change in Finder API that changed Radio Facets to Select
Facets old URLs with preselected options were broken. This fixes it by
using a setter method to cast the value to an array for select facets.
This allows the name of the old radio facet in the URL to be correctly
parsed as a select facet.