Email signup links for content stored Finders
In order to for users to signup for alerts from Finders, we need to
display a link to the signup page. This commit parses the response from
the content store and pulls out the finder_email_signup field in the
Finder initialization if it's present. The nil check happens in the
view, so we assign it as a nil value if the field isn't in the response.

This commit also adds the mail icon (along with a retina version) and
the CSS required to display this.