2 digit years that can be in the past
With the fix for Ruby's bananas YY/MM/DD date format guess, any 2 digit
year would be assumed to be in the 21st century. This commit allows
Users to give 2 digits and get a year in the 20th century by checking if
the year given is in any of the next 50 and if not, assumes it's in the
previous 50.