Beta label for certain Finders
As we launch Finders which may be considered in Beta, we want a way to
display this on a per finder basis. We set this in the metadata in
Finder API, which gets sent the content store as part of the details
hash.

This commit sets this on the Finder object and displays the beta label
if this is true.