Fix contrast and border on iOS
It appears safari on ios8 respects border/background-color settings, which means that the white+no border were rendering these inputs practically invisible on ios8

This commit adds a class to both of the input[type=text] so that we can appropriately scope the white background/grey border to text inputs only. I've used a class instead of input[type=text] so that it works in ie6 too.

With this added we no longer need to overwrite the width and the border which means ios does the default contrasty thing that we want.