Test suite for new date parser
Bin all the old tests, a new approach for this is that we have a whole load of dates that we have seen users entering (through analytics) and we want to support the most common of these.

We have a bit of meta-programming to run through all of these dates. Meta-programming is a little complex (for me, at least), but does allow us to have a test for each date without having to actually write a test for each date. Having one test per date means we have one failing test per failing date, so it's easier to assess the health of our code as we can say "7 out of 10 tests are failing" rather than "our test is failing for an unknown number of dates because it fails for one date and then doesn't run the rest"

The second test stubs takes a date with unspecified year and makes sure the date we assume is this year by stubbing Time.