Re-write date parsing
User research showed us that some users were using a lot broader date types than we were supporting (specifically, just entering a year, no month, no day) We looked into this a little deeper and found there were a couple of date formats we weren't supporting that people were entering.

Now we have a full test suite for all the dates we want to accept, we can add the code required to make it pass. Date parsing is largely handled well by the Chronic gem, with three notable exceptions:

Chronic parses single years ie `2008` as a time: `20:08`. We never expect users to enter a time into a date field on a finder so we are going to assume that four digits is always a year and so convert it to 01/01/yyyy

Chronic doesn't handle eight digits without delimiters. Since eight digits is an unambiguous date so long as you're assuming ddmmyyyy order we can guess at this by converting dddddddd into dd/dd/dddd. The 'so long as you're assuming' here is based on the analytics for this page which show us that a small number of users have put in 8 digits in a row which can be interpreted as ddmmyyyy. None have done yyyymmdd.

Finally, chronic doesn't handle spaces or full stops to delimit dates so when we find numeric dates seperated by these we convert them to /'s

Then pass the date to Chronic, and rescue nil, which is what is returned if the date is invalid.