Refactor Finder to be a Presenter
As we move away from using Finder API, we want to store all the
information for a Finder inlcuding the schema in the Content Store.

This commit is the start of the work that will allow this to happen.
By moving the presentation logic of a Finder from the model to a
Presenter, we follow a more standard Rails pattern. This also allows to
get rid of the horrible `Finder#get` method which fetches the schema
from Finder API.