Refactor facets to use ContentStore response
Now that Finder Schemas are inside the details hash in the response
returned from the ContentStore, we can refactor the facets used and the
Parser.

This also means we can remove the OpenStruct option generation from the
parsers, since all the options are OpenStructs.

While doing this work, I discovered a lot of leftover code that isn't
used anymore such as the `SelectFacet#include_blank` and the methods
related to building native Select boxes so I removed it also.

This commit also contains the updates to the tests.