Remove lists
Back at the beginning of this project, I used lists for these checkboxes because it was thought that organisations would want nested checkboxes within other checkboxes. This has not happened, so I am removing the functionality to have nesting.

First step is removing the lists and replacing the ul with a div, as some of the height calculations that the javascript is doing needs an inner element to the checkbox container.