Truncate after first sentence
After some discussion, truncating after the first sentence seems like
the best option.

This commit does this by gsub everything after a fullstop followed by a
space and an uppercase letter. This functionality can't be achieved with
the normal truncate helper because it works by finding the separator
before the limit, but we wanted it to find the first separator after the
limit.

This way we have a more natural text break, which should hopefully get
around the need for a hard limit.