Finder fronted find or create subscriber list
Instead of relying on finder-api to ensure that all possible subscriber
lists are create for it, finder fronted is now responsible for creating
the list if necessary.

The main addition to make this work is the ability to create a title
for the subscription from the chosen options, and the signup content
item. How the titles should be created is documented at
https://etherpad.mozilla.org/uK2bBcLupr