Make finder#show more agnostic
As we move Finders to being published by multiple Apps we want to make
the Frontend more agnostic to what information it recieves. This commit
does that by specifying that if a value isn't returned in the response
it is false. This happens for the beta flag, beta message and email
signup flag.