Refactor facets to filters with metadata
We can now use the facets to generate the metadata apart from certain
metadata keys which not have a corresponding filter so we need to
introduce a distinction.

This commit does a fair bit of refactoring around this as well as
introducing a `filterable?` method to the existing facet types. I've
also exposed methods for `Finder#filters` and `Finder#metadata` to
access these more easily.

I've also replace the existing Facet class with one which is
unfilterable and all the other facets are subclasses from
FilterableFacet which subclasses from Facet itself.