Handle finders with deep paths
The code was assuming that the finder would always be at /slug, when in the case
of Contact finders, they will be at /government/organisations/org-slug/contact.

The "slug" is used to provide the action attribute to the form, which is then
used by live_search.js to provide AJAX searches. The gsub was mangling deep
paths and so breaking AJAX searches. It was turning:
/government/organisations/org-slug/contact

into:
governmentorganisationsorg-slugcontact