Fix broken markdown links in README
Wrapping a markdown link in code quotes will treat the link brackets as code, and not turn them to links. You can do it the other way round, with code quotes inside the link brackets, but none of the other repo links do that, so I just removed it and fixed the links.