Use Page Title component
Using the standard title component simplies the header, but
still requires limiting the max width by floating. Which
requires clearing on .filtering, as we're floating the
header itself (not the h1 within in, when the header cleared).

This isn't a great pattern, especially as we expect all page
titles to have an effective max width.

Also simplies the margins applied to the header itself, by
always applying a consistent margin.