Allow Finders to have a summmary
For the Policies as a Finder, we want to be able to have a bit of text
to explain what the Finder is. This commit adds the ability for the
Content Item to have a summary and for it to be rendered as Govspeak
using the Govspeak component. Also adding a section to the Finder
Content Item readme about the summary field.