Serve an atom view
This commit adds an atom version of the Finder show page with the feed
entries being the results. It also includes a Presenter for the Atom
feed pads out the FinderController spec to test the show actions when a
content item exists.

In order to provide the correct date this commit also returns the
`last_update` field from Rummager. It uses this from the first entry in
the results array for the feed `updated_at` and on each entry.