Update atom feed auto discovery link on filter
With the autodiscovery link for the atom feed, we want to change it when
the User updates the filter on the page. This commit does that by
modifying the liveSearch code to replace the URL in the head with the
one linked to from the `atom_url` in the response.