Add a government navigation for government finders
This commit adds a render call for the `government_navigation` component
in static. This is only called if the Finder's `base_path` begins with
`/government`. If it does, it also then extracts the second part of the
`base_path` and uses this to decide which link should get a class of
active added.