Extract is_historic and government_name from attrs
This commit is the first bit of work needed in order to display the
government name for documents which were published under a previous
government. In this commit I have modified the `Document` model to
get `is_historic` and `government_name` from the document attrs and
reformatted the `attrs.except` section to be on multiple lines for
readability.

These two bits of metadata are extracted from the normal metadata
pattern because we only want to display `government_name` if
`is_historic` is true and it needs to be displayed on a newline (and not
inline with other metadata).