Add the possibility to add a logo on a finder
European Structural and Investment funds must comply with EU
legislations and they therefore need to display the EU logo.

Quote from the legislation: "the Union emblem and the reference to the
Union shall be visible, when landing on the website, inside the
viewing area of a digital device, without requiring a user to scroll
down the page".