Only use supported metadata fields for a result
Excluding the always-returned-fields (like public timestamp) from
being considered as metadata attrs meant a schema defining
`public_timestamp` for `display_as_result_metadata` would not be
displayed as a metadata.

Rather than excluding the always-returned-fields, only include
the keys defined as for display. This will exclude fields like
`es_score` which are returned for all results, and were previously
included in attrs (but not displayed at all). If we did want to
use `es_score` in the future then it it'd be better to set it on
document directly, as we do for other always-returned-fields.

Add scenario for testing policy metadata, using separate policy schema
and results. The rummager result is done in a slightly hacky way, as
the test helpers assume a mosw schema, and refactoring it would be
quite a bit of work compared to this commit. I'd like to come back
and refactor it to be more generic.

Note; the `return_fields` are de-duped, as public_timestamp will
appear twice, once from base fields and once as a metadata field.
This doesn't cause any problems in the rummager call, but normalising
it is more intuitive and makes tests easier to stub.