Support result metadata strings, not just hashes
Most finders have object-type fields, like `case_state`, where
one the `label` key contains the user-presentable value for the
field.

For Editions, which the policy finders display, some of the fields
that need to be displayed, like `document_type` are string fields
and calling fetch will cause an error, silently ignoring the field.

Allow a field to be either a string or an object to support both.
Switch behaviour on whether fetch is callable, this seemed slightly
cleaner than testing if it was an instance of hash.