Handle fields where the key for display isn't 'label'
Some result fields are objects, rather than string, like
`market_sector` or `report_type`, and one of the keys for
the object will contain the value that should be displayed
to the user.

For specialist-publisher finders (maib, cma, etc) this key is
`label` by convention.

For policy finders we want to show the organisations for a
result, and for the organisations object the key for the
value is `title`, not `label`, as policy documents have
been in rummager in this format for a long time.

This adds some field-aware logic to `FinderPresenter` so it
can know the right display key to use. Originally i tried
adding this to the schema as a facet option, but that added
complexity and wouldn't scale for more complex field display
(like acronym, which we want to use for Orgs in the future).

This keeps the display logic localised to finders for now.

Longer term we should standardise on object field keys in
rummager, so we could rely on something like `label` being
consistent, but thats a long term thing and outside the scope
of this work.