Short-term hack to set default policy date filter
As a user looking at a policy finder
I need to initially see information from the current government only
So that I am not confused by old content that won't be as relevant

https://trello.com/c/sRi6RXKd/118-default-date-view-on-finder-frontend-policy-frontend
.
This this will be used for a few weeks post-election and should be
completely removed afterewards. It only applies to a policy finders, (eg
/government/policies/benefits-reform, but not the finder of policies, eg
/government/policies nor any other finders, eg, /cma-cases)

This will not show documents-related-to-policy published under the previous
government, though they can been seen by removing/changing the published
after date in the finder UI.

This can be removed once there are doucments-related-to-policy published
by the new government. It's intended to be reverted and entirely removed
at that point.

`date_new_government_formed` will need updating to be the day the new
government forms, if it is not the day after polling day.