Disable hacked default policy date fillter
But leave the code in place, so we can use it if we want to.

We were going to set this to the date the government is formed, but
have decided to show all content together for now, and if there is
confusion we can use this.

Regardless of whether we use it or not, it will be removed as part
of the election cleanup work.

Original PR:

https://github.com/alphagov/finder-frontend/pull/192