Quick, hacky, version of faceted filtering
Values for filter facets get fetched from rummager.

This is just enough to get a working prototype together that we can
explore real data with, and possibly user test.

It would need significant cleaning up, probably re-writing, and
testing before it would be shippable.

This requires a couple of steps:

 - not dying on empty allowed_values when constructing the finder
   to make the call to rummager, eg, whitelisting user values,
   which requires constructing the facet objects
 - parsing the rummager results for facets into something usable,
   this is quite hacky, especially around the keys to use for
   'display names'. can't use the string vs hash trick used in
   document metadata, as a string field seems to come back faceted
   as a hash with the string in `slug`
 - inject the faceting results into the allowed_values of facets

All of this is badly shaped vs the rest of finder frontend, but
shows the dependencies at least.