Pass base_path to Airbrake when triggering
When Rummager returns data which doesn't have the key that we need,
Finder Frontend triggers an error with Airbrake. Because this happens in
the model it doesn't have access to the Rack env. In order to get more
information, like what page this is happening on, I've changed this to
pass the `base_path` of the Finder to Airbrake so we at least know what
page this is happening on, but it would be good to refactor it to get
the whole Rack env in future.