Stop collapsed checkbox container using 100% width
For some reason sometimes the collapsed checkbox container takes 100%
width which causes some very weird behavior. It is most strange because
of the absolute positioning it appears over all the other content in the
finder. This forces it to have a small width when it is positioned off
the page and stops it interfering with the content a user can see.