Reorder `FindDocuments` private methods
`#rummager_api` is the first private method refered to by `#call`, so
move it to the top of the list. This moves it out of the way of the
param massaging stuff which is about to get merged into this class -- ie
this is to reduce unrelated churn in a future commit.

`#base_return_fields` is referenced by `#return_fields`, so is moved to
appear beneath it.

fixup 6f0690c