Stop passing `finder` into `FindDocuments`
Pass in the two bits of the finder that are needed instead of the finder
object. This is part of trying to reduce the number of dependencies on
`FinderPresenter` so we can separate the data fetching and massaging
from the presentation.

As part of the refactor this sets the default value for `default_order`
in `#initialize` instead of at the point of use.