Disable atom feed if Finder has default_order
When a Finder has a default_order we want to disable the atom feed from
appearing. The current behaviour of a Finder having an atom feed is
unchanged unless the Finder has a default_order. This commit changes the
atom_url method to return nil if the Finder has a default order, hides
the atom_url link if that is nil in the hash and only responds to the
atom format if the atom_url method isn't nil.