Add `FinderApi`
This provides a `#fetch` method which takes a base_path for the finder
and the user params and gets the finder from content store and the
search results from rummager, and combines them together into a single
response.

It's using `FinderPresenter` to provide the parameter parsing logic,
which is the next thing that needs to be fixed.