Refactor SearchQueryBuilder internals
This class produces a bunch of partial queries in the form of hashes and
returns the combination of those as its output.

To make it more obvious that's what's happening, and to make everything
a bit shallower and easier to dig into, have the `#call` method pull
together all of those partial queries and combine them using `#merge`.

Ending all the method names with `_query` helps to make the purpose of
each method a bit clearer, and reduces the confusion between the
`params` instance variable and the methods building the query.

Also rearranges a couple of methods to stick to the pattern of methods
only calling other methods that are defined beneath them.