Use `Array#+` instead of `Array#concat`
`#concat` mutates the object it's called on. In this case it doesn't
have any negative side effects, because the `base_return_fields` array
is regenerated each time the method is called, but it could lead to
pretty surprising behaviour if the implementation of
`base_return_fields` changed.

If in doubt, avoid mutation.