Remove unused `with_selected_values` method
Nothing uses this method except the tests.

Once this method is removed nothing calls `#selected_value` on a facet,
so the definition in `Facet` can be removed and the definition in
`SelectFacet` can be rescoped to a private method.

Add another test case to `#sentence_fragment`, which is the only public
method that calls the now private `#selected_values` method.