Fix filtering tests
Because static components aren't (currently) rendered in the tests,
this test that uses checkboxes fails as there aren't any. This commit
replaces the steps to find the checkbox and click the "filter these
results" button with call to the url with the filtered results, which
is an equivalent action.

This test removes a side effect of the original test which was checking
for the presence checkboxes on the page and that clicking them and then
filtering updates the result set to a filtered result set.