Add cucumber scenario for date filtering
There's a scenario for filtering documents by metadata, but it doesn't
cover date filtering.

This scenario uses the example cma cases finder from
govuk-content-schemas, fills in a date filter field, submits the form
and checks that the results appear as expected.

Without this feature there is no end to end test for date filtering
which makes refactoring the way the app is architected riskier.