Use FacetQueryBuilder
This commit modifies the FinderAPI and SearchQueryBuiler to use the new
FacetQueryBuiler when querying Rummager in order to get the filterable
facets back which have no allowed_values.

The SearchQueryBuiler then augments the content_item with the results,
the total of the results and the new allowed_values for any filterable
facets which don't have any `allowed_values` present in the content
item.