Explicitly set start offset for rummager calls
`start` is a 0-indexed offset for rummager results (individual
results, not pages).

We're not using it at the moment, but will be when adding
pagination shortly. This makes `start` explicit, even if it is 0,
and explicit is good, especially for constructing reliable URLs
for stubbing requests in tests.

This commit is separate so the ones adding pagination are cleaner.