AJAX Prev and Next page links
This commit updates the Prev and Next links to be modified through AJAX
in order to keep them current with the filters being applied to the
Finder. In order to do this, the component is rendered within the
`ResultSetPresenter` and sent back as HTML within the hash for the
Mustache.