Updating history mode feature test
With pagination now being present in the example content item for a
Policy we need to update this assertion to deal with later changes
involving pagination which reuse the same stubbing.

The change now checks for 5 historical items on the first page as the
Policy in the test paginates at 10 documents.