Join date facet string correctly
When a User enters a from and a to date, it currently joins them as
`"#{from} or #{to}"`. This is incorrect and they should be joined with
`and`. This commit switches this based on the fragment type added
previously and adds a test to check that they are joined correctly.