Change all instances of unified_search to search
This commit changes all instances of unified_search to search when using the GDS API Adapters to access rummager. This provides consistency between the internal and external APIs in terms of naming.

jasmine-rails (a JavaScript testing gem) is also updated as a side effect of phantomjs being updated together with gds-api-adapters.