Default public_timestamp to nil when not present
We are moving old "finder-like" pages over to `finder-frontend`. One of
these finders is the people finder from Whitehall.

The Person records in Whitehall don't have a `public_timestamp` field
because they are not an edition. This means that we don't send that
information to Rummager. When `finder-frontend` fetches those records
from rummager and tries to fetch the `public_timestamp`, the
initialization of a `Document` fails.

This commit makes sure we default to `nil` when fetching the
`public_timestamp`. The reason `nil` is fine is because the default
ordering on people is by `title` (i.e. the minister's names) rather than
by `public_timestamp`.