Hide and report blank facet values
Currently, visiting a finder with a URL that filters on a facet value that is not synced between rummager and the publishing-api results in a blank facet value being displayed and checked.

This commit excludes facet values with no title from being displayed, and also raises an Airbrake error so that these problems are logged to be fixed.