Ensure Jenkins build parameters are set
Jenkins build parameters do not seem to be set on the very first build
of a new branch. env.PARAM_NAME returns null, and referring to them
directly by PARAM_NAME throws groovy.lang.MissingPropertyException.

This is a very hacky solution - it checks for null parameters and makes
sure that they are set to the defaults if so - but I haven't been able
to find a better solution.