Change handling of email alert titles
This commit splits the `title`, `short_name` and `description` attributes so that they can be separately generated, whereas before `title` was also used for `short_name` and `description` was empty when passed to the `email-alert-api`.

Since `title` and `short_name` have a 255 character limit enforced by GovDelivery, this commit adds logic to keep them under this limit: only the name of the finder and the number of topics chosen are shown, followed by an MD5 hash of all the topic names (for `title` only to keep it unique). `description` stores all the names of the chosen topics, which was previously stored in `title` and `short_name`.