Fix intermittent test failure on CI
This hopefully fixes intermittent test failures on CI like this one:

https://ci.integration.publishing.service.gov.uk/job/finder-frontend/job
/deployed-to-production/592/console

These seem to be caused by not having stubbed
`stub_shared_component_locales`, but I haven't been able to reproduce
this locally. Even when running the tests with the same seed.

This adds the stubbing for all tests so that we can be sure it's not an
ordering issue.