Track clicks on links in finders
This commit adds data attributes to links in all finders that will help
us track how those links are being used by our users.

There is no concept of index in mustache, which means I had to change
`documents` to include the index of the documents so that it can be
added to the data attributes to report the position of the link.

Trello: https://trello.com/c/vRtpDbnk/56-investigate-getting-nav-beta-analytics-tracking-onto-finder-template