Use hash accessors on gds-api-adapter objects
Since v37.0.0 the data that comes back from a gds api adapters in the
response has acted like a hash, not an openstruct which means we can't
do:

    content_item.details.facets

but have to do:

    content_item['details']['facets']

instead.  It's almost certain we've not picked up all the places in the
app where we are still trying to call a method on a hash, but these are
the obvious ones.