Use hashes instead of open structs for sentence fragments
This data doesn't come from the api response, but it makes everything
consistent as it's all a hash now, and also avoids using OpenStruct which
has poor performance characteristics.

Arguably a hash is less easy to work with, so once this transition is
complete we may want to look into creating some simple value objects and
move back towards method invocation instead of hash access.  However, it'd
be good to do this across the board and keep things consistent.