Use rails 5.1 style parameter permissions for search parameters
This avoids a deprecation warning when running the cucumber features about
calling `.symbolize_parameters` which will no longer be supported in rails
5.1.  We have to add a config option to tell rails 5.0 to use the new 5.1
behaviour when calling `.to_h` otherwise we get another deprecation
warning.  This config option is

    action_controller.raise_on_unfiltered_parameters = true

And is the default behaviour in rails 5.1 so will pave the way for us to
upgrade in the future.

Note that we permit the `filter_organisations` and `filter_manual` facets
to be provided as single values, or as array values.  It's not clear from
the code if only `filter_organisations` should have this behaviour though.