Add note about SelectFacet#options
This is the only symbol-keyed has data structure in the app at the moment,
everything else is string-keyed.  This is symbol-keyed because it is
passed as an argument into a govuk_component partial that expects symbol
keys.  I think it's worth highlighting this because everything else in
the app is a string-keyed hash because it comes out of an api response
(or is added by us to a hash that comes out of an api response and we
don't want to mix our key types).