Fix email alert subscriptions view after openstruct -> hash change
In #342 we changed this app to use the newest gds-api-adapters which meant
changing all our data from openstructs to hashes.  We missed some old
references in the email alert subscriptions views that still used method
invocation to get data instead of hash accessors.  To help us avoid this
in the future we add a happy-path test case for the email alert
subscriptions controller and ask it to render the views.  We also make
the fixture for the signup content item more realistic, with some email
signup choices and extra details.