Remove doubles from email_alert controller and lib specs
We prefer to use the real instances, and use the http request stubbing
methods provided by the gds-api-adapters gem.  This gives us some
confidence that our tests exercise the code under test in a more realistic
setting.  Particularly as there don't appear to be any integration or
acceptance tests for signing up to an email alert.