Stop treating email_alert_api responses like openstructs
Our tests used doubles that had methods instead of using the real api
adapters with stubbed http responses so they didn't alert us to these
remaining method invocations.  In the previous commits we changed the
tests so they stopped using doubles and this finally gave us a broken
build that prompts us to fix this.