Handle invalid taxons params gracefully
The path for the taxon to filter results by is passed
as a request param. We need to fetch the taxon by path
from the content store which will return an not found
error if an item at that path doesn't exist. We should
just omit the taxon filter from the search in this case.