Close large select facets by default
Select facets are collpased by default if they are not the first facet of this
type on the page.

We want to embellish this such that large facets are not open by default on finders.

This is in response to this [Zendesk ticket](https://govuk.zendesk.com/agent/tickets/3058159)
noting that https://www.gov.uk/government/organisations/latest has the
organisation facet open by default.  This is fine on desktop, but when viewed
on mobile, the facet gets in the way, and users have to scroll past hundreds
of orgs before getting to the content.

I've checked this out on the [other finders](https://www.gov.uk/api/search.json?filter_content_store_document_type=finder),
and I think it's OK.