Lookup content purpose document types to query
This changes the way advanced search interfaces with the Search
API. Instead of querying by content purpose subgroup or supergroup we
instead lookup the applicable document types and query for them
directly.

This is part of a series of changes required to remove the dependency
on the Search API including content purpose fields in its index. This
will allow us to iterate the content purpose mappings without having
to reindex every item of content.