Change the expectation regarding advanced search 404's
Due to whatever Rails is doing in the controller, it seems set on
rendering some HTML when it actually should be responding with a plain
404.

I can't work out how to make it do something more sensible, so just
change the expectation to look at the status code. This shouldn't make
any difference to end users, as the actual content that users will see
is the standard GOV.UK 404 page.