Present taxon base_path as topic filter param in results
The topic filter param is rewritten by the AdvancedSearchFinderApi in
order to generate a valid Rummager query. This doesn't work when rendering
pagination links as the value is now the topic's content_id and not the
original base_path. So make sure we replace with the original base_path value
before rendering the pagination links.