Fix event tracking on keypress
Previously, we were only firing tracking events when the user submitted their text search by pressing the enter key. However, users may tab away from the field (for example) to search, and this wouldn't fire.

This commit changes the behaviour so events are fired on 'change' instead of on 'keypress'. A change event is always fired, even if the user also presses the enter key. We can't have the event firing on both as then we would get duplicate events in Google Analytics.